Fix for #2
Fixes #2. The event data object is manipulated when a request is passed from the page script to the client script.
Firefox uses the same event object for all event handlers and therefore the changed event data is handed to the message event handler in page.js.
This causes the CLIENT_INTERNAL event to be mistaken for a CLIENT_EXTERNAL event.